+++ part-last/my-board-spring5/src/main/java/org/fp024/domain/ReplyVO.java
@@ -1,22 +1,24 @@
 package org.fp024.domain;
 
+import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+import org.hibernate.annotations.DynamicInsert;
 
 @NoArgsConstructor // TODO: 컴파일 오류 방지
 @ToString
 @Getter
 @Setter // TODO: 컴파일 오류 방지
 @Entity
+@DynamicInsert // INSERT 할때 null 인 경우 쿼리에 포함시키지 않음.
 @Table(name = "tbl_reply")
 public class ReplyVO {
   @Id
