+++ part-last/my-board-spring5/src/main/resources/database.properties
@@ -1,5 +1,5 @@
 jdbc.driver=org.hsqldb.jdbc.JDBCDriver
-#jdbc.url=jdbc:hsqldb:mem:my-board
-jdbc.url=jdbc:hsqldb:hsql://hsqldb-host:9001/my_board
+jdbc.url=jdbc:hsqldb:mem:my-board
+#jdbc.url=jdbc:hsqldb:hsql://hsqldb-host:9001/my_board
 jdbc.username=SA
 jdbc.password=
