+++ part-last/my-board-spring5/src/test/java/org/fp024/repository/jpa/BoardAttachRepositoryTest.java
@@ -77,7 +77,8 @@ class BoardAttachRepositoryTest {
   void delete() {
     Long bno = 1L;
     long deleteRowCount = repository.deleteByBno(bno);
-    assertThat(deleteRowCount).isEqualTo(1);
+    LOGGER.info("### deleteRowCount: {}", deleteRowCount);
+    assertThat(deleteRowCount).isGreaterThan(0);
   }
 
   /*
