+++ setenv.properties
@@ -1,4 +1,4 @@
 JAVA_HOME=/home/fp024/JDK/17
 SCOUTER_JAVA_AGENT_CONF=scouter/conf/board-webapp.conf
 SCOUTER_JAVA_AGENT_LIB=/home/fp024/apps/scouter/latest/agent.java/scouter.agent.jar
-LATEST_PROJECT_HOME=part07/jex06-board
+LATEST_PROJECT_HOME=part-last/my-board-spring5
